Refactor into multiple files and add config options
