+++ src/main/java/de/hysky/skyblocker/config/categories/DungeonsCategory.java
@@ -155,11 +155,25 @@ public class DungeonsCategory {
                         .option(Option.<Boolean>createBuilder()
                                 .name(Text.translatable("skyblocker.config.dungeons.puzzle.solveWaterboard"))
                                 .description(OptionDescription.of(Text.translatable("skyblocker.config.dungeons.puzzle.solveWaterboard.@Tooltip")))
-                                .binding(defaults.dungeons.puzzleSolvers.solveWaterboard,
-                                        () -> config.dungeons.puzzleSolvers.solveWaterboard,
-                                        newValue -> config.dungeons.puzzleSolvers.solveWaterboard = newValue)
+                                .binding(defaults.dungeons.puzzleSolvers.waterboardOneFlow,
+                                        () -> config.dungeons.puzzleSolvers.waterboardOneFlow,
+                                        newValue -> config.dungeons.puzzleSolvers.waterboardOneFlow = newValue)
                                 .controller(ConfigUtils::createBooleanController)
                                 .build())
+						.option(Option.<Boolean>createBuilder()
+								.name(Text.translatable("skyblocker.config.dungeons.puzzle.previewWaterPath"))
+								.binding(defaults.dungeons.puzzleSolvers.previewWaterPath,
+										() -> config.dungeons.puzzleSolvers.previewWaterPath,
+										newValue -> config.dungeons.puzzleSolvers.previewWaterPath = newValue)
+								.controller(ConfigUtils::createBooleanController)
+								.build())
+						.option(Option.<Boolean>createBuilder()
+								.name(Text.translatable("skyblocker.config.dungeons.puzzle.previewLeverEffects"))
+								.binding(defaults.dungeons.puzzleSolvers.previewLeverEffects,
+										() -> config.dungeons.puzzleSolvers.previewLeverEffects,
+										newValue -> config.dungeons.puzzleSolvers.previewLeverEffects = newValue)
+								.controller(ConfigUtils::createBooleanController)
+								.build())
                         .option(Option.<Boolean>createBuilder()
                                 .name(Text.translatable("skyblocker.config.dungeons.puzzle.blazeSolver"))
                                 .description(OptionDescription.of(Text.translatable("skyblocker.config.dungeons.puzzle.blazeSolver.@Tooltip")))
