+++ src/main/java/de/hysky/skyblocker/skyblock/dungeon/puzzle/DungeonPuzzle.java
@@ -1,21 +1,13 @@
 package de.hysky.skyblocker.skyblock.dungeon.puzzle;
 
-import com.mojang.brigadier.Command;
-import de.hysky.skyblocker.SkyblockerMod;
 import de.hysky.skyblocker.events.DungeonEvents;
-import de.hysky.skyblocker.skyblock.dungeon.secrets.DungeonManager;
-import de.hysky.skyblocker.skyblock.dungeon.secrets.Room;
-import de.hysky.skyblocker.utils.Constants;
 import de.hysky.skyblocker.utils.Resettable;
 import de.hysky.skyblocker.utils.Tickable;
 import de.hysky.skyblocker.utils.render.Renderable;
-import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Set;
-
-import static net.fabricmc.fabric.api.client.command.v2.ClientCommandManager.literal;
 
 public abstract class DungeonPuzzle implements Tickable, Renderable, Resettable {
     protected final String puzzleName;
@@ -36,17 +28,6 @@ public abstract class DungeonPuzzle implements Tickable, Renderable, Resettable 
                 shouldSolve = true;
             }
         });
-        ClientCommandRegistrationCallback.EVENT.register((dispatcher, registryAccess) -> dispatcher.register(literal(SkyblockerMod.NAMESPACE).then(literal("dungeons").then(literal("puzzle").then(literal(puzzleName).then(literal("solve").executes(context -> {
-            Room currentRoom = DungeonManager.getCurrentRoom();
-            if (currentRoom != null) {
-                reset();
-                currentRoom.addSubProcess(this);
-                context.getSource().sendFeedback(Constants.PREFIX.get().append("§aSolving " + puzzleName + " puzzle in the current room."));
-            } else {
-                context.getSource().sendError(Constants.PREFIX.get().append("§cCurrent room is null."));
-            }
-            return Command.SINGLE_SUCCESS;
-        })))))));
         ClientPlayConnectionEvents.JOIN.register(this);
     }
